% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marini.R
\name{busca_marini}
\alias{busca_marini}
\title{Buscar nos textos do Marini}
\usage{
busca_marini(termo)
}
\arguments{
\item{termo}{Termo buscado no titulo}
}
\value{
Uma \link{[tibble]tibble}
}
\description{
Buscar nos textos do Marini
}
\examples{
busca_marini("Cuba")
}
