% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{escritos}
\alias{escritos}
\title{Criar links completos}
\usage{
escritos(final)
}
\arguments{
\item{final}{link relativo}
}
\description{
Criar links completos
}
\examples{
escritos("015_indice_general_marini.html")
}
