% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{estruturar_indice}
\alias{estruturar_indice}
\title{Criar tibble com todos textos do portal}
\usage{
estruturar_indice()
}
\description{
Criar tibble com todos textos do portal
}
\examples{
# Deve ser usada apenas internamente
}
