% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{estruturar_texto}
\alias{estruturar_texto}
\title{Estruturar dados de texto do portal}
\usage{
estruturar_texto(link)
}
\arguments{
\item{link}{Link de um texto no}
}
\value{
Uma \code{\link[tibble]{tibble}} com dados do texto que
  se encontra no \code{link}.
}
\description{
Estruturar dados de texto do portal
}
