% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marini.R
\name{nuvem}
\alias{nuvem}
\title{Gerar nuvem de palavras}
\usage{
nuvem(df)
}
\arguments{
\item{df}{Um \code{data.frame}}
}
\value{
Uma nuvem de palavras
}
\description{
Gerar nuvem de palavras
}
\examples{
busca_marini("Cuba") \%>\%
  nuvem()
}
