% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pegar_id}
\alias{pegar_id}
\title{Extrair ID do texto usado no portal}
\usage{
pegar_id(link)
}
\arguments{
\item{link}{link do texto no portal}
}
\description{
Extrair ID do texto usado no portal
}
\examples{
escritos("015_indice_general_marini.html") \%>\%
  pegar_id()
}
